[Fix] 버그 수정 및 모델 데이터 타입 변경

+++ Assets/KwakSeongDae/Scripts/ResourceUIController.cs
@@ -44,12 +44,12 @@ public class ResourceUIController : MonoBehaviour
         }
     }
 
-    void UpdateMoney(int newMoney)
+    void UpdateMoney(long newMoney)
     {
         moneyText?.SetText(newMoney.ToString());
     }
 
-    void UpdateJewel(int newJewel)
+    void UpdateJewel(long newJewel)
     {
         jewelText?.SetText(newJewel.ToString());
     }
